+++ src/views/DestinationDetailView.vue
@@ -13,6 +13,7 @@ import ItineraryBuilder from '../components/ItineraryBuilder.vue';
 import PackingChecklist from '../components/PackingChecklist.vue';
 import TravelMapVisualizer from '../components/TravelMapVisualizer.vue';
 import WeatherClimateWidget from '../components/WeatherClimateWidget.vue';
+import UserTravelStories from '../components/UserTravelStories.vue';
 
 const route = useRoute();
 const travelStore = useTravelStore();
@@ -56,6 +57,7 @@ const relatedPosts = computed(() => {
         <WeatherClimateWidget :destination-id="destination.id" class="margin-top-calc" />
         <BudgetCalculator :destination-id="destination.id" class="margin-top-calc" />
         <ItineraryBuilder :destination-id="destination.id" class="margin-top-calc" />
+        <UserTravelStories :destination-id="destination.id" class="margin-top-calc" />
         <TravelMapVisualizer :highlight-dest-id="destination.id" class="margin-top-calc" />
         <PackingChecklist class="margin-top-calc" />
       </div>